Add ability to selectively re-compute liveness in vasm-graph-color
Summary:
Instead of recalculating liveness from scratch whenever we need it,
add the ability to selectively re-compute it (on a per-block
basis). Store the data that the liveness calculation uses (uses, defs,
and gen sets) in the allocator state. If a bitset of blocks changed
from last calculation is provided, selectively re-compute the
liveness, re-using the old information where-ever possible.

Right now vasm-graph-color doesn't provide the changed block list, but
it will in a later diff.

Reviewed By: mofarrell

Differential Revision: D18849432

fbshipit-source-id: f843da887042fc17b710e986813c27b5f553988a