Track which blocks have changed during spilling in vasm-graph-color
Summary:
During spilling, track which blocks have been modified and feed this
information into compute_liveness() and restoreSSA(). The only
remaining case we can't track is when DCE removes a dead spill. In
that case, we mark all blocks as modified. The next diff will fix that
(and make the DCE cheaper).

Reviewed By: mofarrell

Differential Revision: D18849880

fbshipit-source-id: 54e38929a9364dbdb6027572c48f309d8ca56940