Implement specialized dead spill elimination in vasm-graph-color
Summary:
Up until now, we've been relying on the standard DCE pass to remove
any dead spills. This works, but is somewhat expensive. We can get
better results by implementing a special dead spill removal pass. We
can leverage our set of changed blocks to avoid visiting many of the
blocks. In addition, we can track exactly which blocks we changed,
which speeds up the liveness recalculation we have to do afterwards.

Reviewed By: mofarrell

Differential Revision: D18864205

fbshipit-source-id: 5f189671eab649b564ba5eb5be475264daf7f72f