Speed-up spill weight calculation in vasm-graph-color
Summary:
Now that we have per-block use sets, when calculating spill weights,
we can skip over blocks which we know won't use any of the interesting
Vregs. This has one semantic change. Previously we would only count
non-copy instructions for calculating the weight. This is difficult to
do when skipping the entire block, so now count all instructions. I
don't think this difference is meaningful for performance.

Reviewed By: mofarrell

Differential Revision: D18867042

fbshipit-source-id: 39cd238e64197b660336dfd8af526b085f8bcfea