Speed-up liveness calculation in vasm-graph-color
Summary:
Rather than looping over the use/def sets for each instruction, we can
leverage the bitset representation and use bitwise operations
instead. The only complication is we want to keep flag registers out
of the sets. Fix this by keeping a VregSet of all flag registers which
we can then subtract out as needed.

Reviewed By: mofarrell

Differential Revision: D18868496

fbshipit-source-id: 79495fa62a58f69a2b570a1640a13766db177072