Fix incorrect pops when const folding builtins in HHBBC
Summary:
Some builtins can take Uninit values as parameters
(representing an unsupplied optional parameter). If we const fold such
builtins, we'll run into issues because we always try to pop the (now
unused) arguments using PopC. Instead check the known type before
emitting the pop to know whether to emit a PopC or a PopU. This won't
work if we have a Cell as parameter, but they can't happen right now
and there's no way to know which pop to emit in that case. This const
folding issue becomes a problem when I start trying to optimize
Shapes::idx in HHBBC.

Reviewed By: markw65

Differential Revision: D18921345

fbshipit-source-id: 575a9b54286efe83a6604c2040009eb4b51e1b38