record queue time
Summary:
For messages to hh_server, we record (1) what time the message arrived at stdin from VSCode/Nuclide, (2) what time hh_server got around to actually start handling the message, (3) what time the message handling was finished and a result delivered to the client.

The mechanism for (2) in hh_server is first that clientLsp doesn't pick anything up from its incoming jsonrpc queue until it's finished with previous work, and second that the message sits in a blocking call to send data over the pipe to hh_server until such time as the pipe is ready to receive it, and third that data sits in the pipe until hh_server's message-loop is ready to read from it and act upon it.

We should also be logging (2) for ClientIdeDaemon. The equivalent there is that a message might be in the pipe to ClientIdeDaemon, or clientIdeDaemon might have picked it up from its pipe and stuck it in the message queue.

This diff adds (2) for ClientIdeDaemon. It marks the message as 'unblocked' at the moment where "handle_message" picks it up and starts acting upon it.

All of the other code is already in place to log this queue time to HackEventLogger and Hh_logger.

NOTE: why am I doing this with a parameter `ref_unblocked_time : float ref` rather than making each method return a tuple/record/parameterized ADT? Honestly, just because this way took fewer lines of code and looked more uniform.

Reviewed By: 2BitSalute

Differential Revision: D18919615

fbshipit-source-id: 125c05ac8831913cb3b48547c061a994e3d20682