count time spent fetching decls
Summary:
We're already logging how many decl cache misses we encountered while typechecking a file. This diff adds up the length of time spent fetching those decls.

I've split up the old function `Deferred_decl.count_and_raise_if_should_defer` into two separate functions: (1) `raise_if_should_defer` called at the beginning of when we encounter a missing decl, (2) `count_decl_cache_miss` called after we've finishig fetching the missing decl, since this second one needs to know how many ms it took us to fetch that decl.

This is a slight change in internals. It means that we will raise without having incremented the counter. I've changed the comparison from `counter > threshold` to `counter >= threshold` accordingly so we'll get exactly the same behavior. But it means that from now on, anyone who logs the `counter` will see it only having counted the decls that were successfully fetched, not the ones that triggered a defer.

(the only place this is examined is in telemetry, and we're not yet using the telemetry, so the change is unimportant.)

Reviewed By: 2BitSalute

Differential Revision: D18928579

fbshipit-source-id: d654fa462bd965311a84a4336985a844f927f700