pass on --verbose flag to clientIdeDaemon
Summary:
An earlier diff introduced the flag `hh_client lsp --verbose`, with the effect of setting the min-log-level so that `Hh_logger.debug` statements get logged as well.

(The intention is for this to cover voluminous data, including per-message data, for help in hack developers to diagnose issues.)

This diff merely passes the flag on to ClientIdeDaemon. The verbosity of Hh_logger needs to be set once per process.

Reviewed By: arxanas

Differential Revision: D18936184

fbshipit-source-id: 3452e1bc49b537c3cd3f6f09ae0af8c9a4da9835