Split Tabstract into individual types
Summary:
Tabstract was introduced by me several years ago to make it easier to deal with the various types that had upper bounds. But now generics store their bounds in the typing_env and newtype/dependent types always set the "optional" upper bound type (defaulting to mixed).

This abstraction doesn't provide much value in the current state of the type checker and requires remembering to set the upper bound for dependent type and newtypes but not for generics. Let's split them out to separate types.

Reviewed By: andrewjkennedy

Differential Revision: D18949263

fbshipit-source-id: f18ec331b903476f719453436423b0df9a331aca