Fix HHBBC DCE logic around PopU2 and PopFrame
Summary:
If HHBBC decides to kill the Cell values on the stack above the
Uninits that a PopU2 or PopFrame will remove, it will simply remove
the Cell values and not change the PopU2 or PopFrame. This means the
PopU2 or PopFrame will pop off the wrong things. To fix this,
introduce a new DCE action called AdjustPop. This action is only for
PopU2 or PopFrame and indicates that the bytecode should be modified
to account for the Cell(s) disappearing. For PopU2 this means turning
it into a PopU, and for PopFrame, either just changing the immediate
or lowering it to a set of PopUs or PopU2s.

Differential Revision: D18922406

fbshipit-source-id: f21a24a765da79bff4406751ddb20a2cb9ec5afd