Initialize spiller initial state faster in vasm-graph-color
Summary:
When initializing the initial spiller state for a block, we look at
all of the predecessors of the block, what they've decided for the
reg/mem state for each Vreg, then invoke the spilling machinery to
decide what Vregs should be in regs or memory as of the entry to the
block. However, if all the predecessors agree on the state of all
live-in Vregs, we can skip all of this. Instead we can just copy the
out state of one of the predecessors (doesn't matter which since
they're all the same) and use it as the initial state.

Reviewed By: mofarrell

Differential Revision: D18873794

fbshipit-source-id: b7e23ed2c0769ffd655bdbe8928d4ee2976959ce