Update hhvm_wrapper now that refs are gone
Summary: This script used two reffy functions: preg_match and manipulation on an array. We replace the first one with inout preg_match_with_matches and the second with a vector. Note that this code modifies the argv global which downstream code relies on; we match this behavior even though it's quite ugly.

Reviewed By: jjgriego

Differential Revision: D18941409

fbshipit-source-id: d64cad2caeb79e3da231fc2156f4d209259ccbc2