Fix incremental naming update exception
Summary:
See the comments in the changed code.

Previously, our code to do a reverse naming table update re-read files from disk to ascertain whether symbols were still in the naming table. This is 1) racey because files can change on disk, and 2) inefficient because we have all the information we need in hand to answer that question anyways.

This diff doesn't handle the fact that technically our naming table update is incorrect in the presence of duplicate names. However, sIDE should function mostly fine in those scenarios, and I haven't seen any reports in the wild of users actually suffering from the results of duplicate names.

Reviewed By: ljw1004

Differential Revision: D18977170

fbshipit-source-id: 06cbc481d5c878025d116157c94723f5c9614ec6