Hoist incremental file update exception handler
Summary:
`ClientIdeIncremental.process_changed_file` should never throw an exception. If it does, the caller should know and choose whether to log/abort/etc. For that purpose, let the exception propagate upwards one frame.

This commit also changes it to log the exception in the regular log files, since that's a very important piece of data.

Reviewed By: ljw1004

Differential Revision: D18977169

fbshipit-source-id: 2c3b3798036498613ad0a0bd659b36b09dc67d13