Wait for file changes to be processed after `write_to_disk` in tests
Summary:
The default behavior in tests should be to wait for the `write_to_disk` call to be processed by the language server before proceeding. Otherwise, sIDE will appear to ignore the file change because it's busy prioritizing IDE requests over file change notifications.

In the event that a test writer cares, they can opt out of this behavior.

Reviewed By: ljw1004

Differential Revision: D18947284

fbshipit-source-id: 05e57020bf67ba6eb4183eb6efcf206740bb052c