Fix some gdb python helper issues
Summary:
Various recent changes have broken hhx again. The tests that were
supposed to catch this were disabled because they sometimes timed
out. The new gdb indexer fixed the timeouts, but broke the tests in
new ways.

This fixes the tests so they work when the new indexer is
disabled. The next diff will fix the indexer.

Reviewed By: oulgen

Differential Revision: D18972036

fbshipit-source-id: 586943093407c78ed52a563e049fc11328f776f2