Fix naming of things in gdb-indexer
Summary:
For reasons that aren't clear, gcc often outputs a declaration of a
thing, properly nested in its namespaces etc, followed by a non-nested
definition, which has a specification pointing to the declaration. The
definition may or may not contain a name at all, but if it does, it's
just the leaf name.

Keep a table mapping declaration offsets to their names, and if we see
a DW_AT_specification, lookup the name there.

This fixes a number of missing symbols, and incorrectly named symbols
in the generated index.

Reviewed By: oulgen

Differential Revision: D18972037

fbshipit-source-id: 1355fe9f83720ffe323973cd35e68c994d155dc0