Stack reporting driven by signals from BPF
Summary:
BPF-based profiling has become a common method of profiling,
but BPF programs are not able to poke into HHVM and interrogate the
stack of async execution. This is a prototype of a workaround that
allows a BPF program to request a report of the HPHP stack by issuing
a POSIX signal.

bpf_send_signal has been recently added to the kernel, and can be used
to issue the signal. A client BPF program can then read the reported
stack by attaching a program to the "hhvm_stack" USDT.

Reference code is provided in a separate diff upstack

Reviewed By: binliu19

Differential Revision: D17631165

fbshipit-source-id: 6de6a2dd4b006a3af869b6c9f1adaad3d22ec950