Ability to save naming table diff independently
Summary:
Also removing the ~update_reverse_entries flag. This flag is meaningful only in the Sandcastle saving state scenario:
- If we create a new saved state based on a previous SQLite-backed naming table, then, when loading this previous saved state, we need not update the reverse naming table entries, as they won't be used.

This flag is confusing, as is indicated by the fact that it's incorrectly set to false in the client IDE code.

We currently don't implement this flow, anyway, and with the separate diff saving, it may not be needed at all.

Reviewed By: ljw1004

Differential Revision: D18723777

fbshipit-source-id: f4c9f6e96680357aefe35f54dee6b30ddd7f4801