Make spilling determination logic more specific in vasm-graph-color
Summary:
Make the logic to determine if a block can potentially spill in
vasm-graph-color more specific. Instead of just having a bit
per-block, instead track it on an instruction by instruction basis.

We can leverage this in the spiller to make it more efficient. We
already skip blocks which cannot trigger a spill. If an instruction
cannot trigger a spill, we can also process it more efficiently. We
can leave the spilled Vregs untouched, and just add the defs of the
instruction to the non-spilled Vreg set. After this, we just have to
remove dead Vregs and we're done. This is significantly more efficient
than the full spiller logic.

Finally, make the spilling determination logic slightly more
efficient. Instead of tracking two liveness sets (one for GPs and one
for SIMDs), just keep a single set. If the size of the set is smaller
than the common limit of the two, there's no need to distinguish the
two reg-classes. If they don't, we can then split them out by
reg-class and do per reg-class checks. Since we don't use SIMDs too
often, this is more efficient.

Reviewed By: mofarrell

Differential Revision: D18875897

fbshipit-source-id: 8e4e0ce8069663591e773b39f3c0cf0786f5e70e