Insert @generated line to files with no manual sections
Summary:
As in title. To make it 100% correct the new line should be within signed section, but that would require creating new per-file generated section in addition to individual sections corresponding to targets, and this is too much rearchitecting for me. I doubt anything can realistically go wrong because of this.

(Note: this ignores all push blocking failures!)

Reviewed By: lukaspiatkowski, dtolnay

Differential Revision: D18837858

fbshipit-source-id: c84fb9c371f5fe6b895cd15e380ccb2d1e374667