log RequestCancelled
Summary:
Oops! Back in Jan 2019 we accidentally stopped logging telemetry for "RequestCancelled" LSP responses. D13514214.

RequestCancelled happens e.g. because a hover request was made a while ago and has been superceded by a newer request and so we might as well not even both handling the original hover request. Another case is that the server is busy initializing, during which time we respond to hover and other requests with RequestCancelled. (We started using RequestCancelled here because it was the only form of LSP response which didn't trigger the console to pop up in Nuclide. That was subsequently relaxed by Nuclide folks, but we still do it).

The intent of that Jan 2019 diff was to stop logging RequestCancelled to `hh --lsp-logname` and stderr. That was a good change because it happened a lot.

But we still need to gather telemetry about all RequestCancelled. We need that (1) as the denominator so we can judge the success rate of message handling, (2) because otherwise we have no insight into which messages were denied due to hh_server initialization.

Differential Revision: D19090174

fbshipit-source-id: 1a9f2161007af52782f654233404cdee82a4d5c3