Detect duplicate notifications when waiting for them in tests
Summary:
When we wait for a message, we check to see if it's in our transcript so far. If not, we read messages until it is.

For requests, we make sure to record the IDs of messages that we've seen before. That way, if we get a duplicate response, we actually wait for both responses. We don't just examine the transcript and see the first response and proceed as if the second request had gotten its response.

There is no such system for notifications. This masks issues where we can't correctly wait for duplicate notifications.

The fix is to record the "transcript ID" of each message we've seen before, rather than just the JSONRPC request ID. Notifications do not have request IDs because they are not requests. Every message has a "transcript ID" because we assign one in our testing framework.

Reviewed By: ljw1004

Differential Revision: D18947285

fbshipit-source-id: 4942c4a1412177e98c36bd3e766083f02bc3485b