Typing check service progress: adjust calculation of completed count
Summary:
We keep the work-in-progress in a hashset. If a worker comes back, and it ends up being that another worker already type checked some of its payload, then we could end up overcounting how many files we successfully rechecked.

Instead, we can increment the completed count by the number of files we remove from the work-in-progress hashset.

Differential Revision: D18934847

fbshipit-source-id: f24a2bd2786324dc976e9ae7db318211c7f29d5e