Add ability to serialize profile data for optimized JITed code
Summary:
This diff adds a basic framework to collect profile data for optimized JITed
code.  This is still disabled and not used for anything yet.  When enabled
though, hhvm will run for a specified number of requests or time after
retranslate-all finishes, and then it can append additional profile data to
the JIT serialized data file.

This diff also updates the test runner to use this new mode with --jit-serialize.
The tests run once more after the retranslate-all and can serialize profile data
collected for optimized code after that extra iteration.

Reviewed By: binliu19

Differential Revision: D18904536

fbshipit-source-id: 9f3f723e543a929feb533274e30105e2ee46828a