Find more symbols for the index
Summary:
static members first appear under a DW_TAG_member declaration, and
then later as a DW_TAG_variable with that member as its specification.

We have to record the name on members so we can correctly report the
variable. This also means we have to recurse on structured types so
that we see the members.

Similarly for static methods.

Reviewed By: ricklavoie

Differential Revision: D19136122

fbshipit-source-id: 3b59bd525c4ee7c08beaccafd1928f3f36726756