Type accesses cleanup: do not thread the expand environment
Summary:
The `ety_env` was needlessly passed around in the local 'env'. Only data
which has to be mutated needs to be passed around this way.

In this diff I make it clear that `ety_env` is simply passed downwards the
call stack, but never changed by functions in Typing_taccess. This pattern is
similar to how `ety_env` is handled during localization.

Reviewed By: dlreeves

Differential Revision: D19091339

fbshipit-source-id: dbe6c984829d824bffe81da92cc8e3a4ae2eaac5