Implement OcamlRep for OsString
Summary:
OsString represents platform-native strings (and is the backing storage for PathBuf). Most OS-provided strings we handle are paths, but a few aren't (like command line arguments), so an implementation of OcamlRep for OsString should be useful.

We can repurpose the PathBuf implementation to expose an impl for OsString. Since PathBuf can be converted to/from OsString without allocating or copying memory, we can implement PathBuf::from_ocamlrep using OsString::from_ocamlrep.

Reviewed By: 2BitSalute

Differential Revision: D19147658

fbshipit-source-id: bd4af6981e2f0abcae2b74b7a3f0b35695508b89