serverless-IDE reports callstacks
Summary:
I was frustrated reading the telemetry to see lots of reecords like "serverless-IDE is stopped/crashed" but with no way to see why it's stopped, short of looking back at each individual transcript and seeing if I could find anything earlier. That's not good enough for when we want to aggregate data. The goal of this diff is to preserve callstacks so we can always directly identify the reason for an LSP-handling failure.

This diff makes sure we never lose the originating callstacks:
1. for an RPC call that failed, we always keep the remote callstack of the failure
2. for when the clientIdeDaemon is in stopped/crashed state, we remember the callstack that took it into that state.

Reviewed By: arxanas

Differential Revision: D19115431

fbshipit-source-id: 45b21029f3c1b63675d5a73dafd85394d60b4c07