Fix bug in printing of like types
Summary:
There was a bug in {D15349710} where we didn't properly handle all the cases when printing like types in Typing_print. This led to likes types showing that they were nullable even if they weren't, if they were a part of a union that contained dynamic and more than one nonnull type.

I switched to using booleans instead of list for the dynamic and null case. I also do not use wildcards in the pattern match so we get verification from the compiler that we handled all cases

Reviewed By: andrewjkennedy

Differential Revision: D19157329

fbshipit-source-id: c1c928392ca0faeb48f8795353b86690fd4a53bd