Return type for hh pseudo functions should be void
Summary: Currently Hack generates a fresh type variable for the return type of functions such as `hh_show`, `hh_show_env`, etc. Typically it then "solves" this type variable to be `nothing`. Eventually, when we can trust types (or use them in linting), this would denote functions that don't return or always throw an exception, so code below the call can be treated as dead. This isn't what we want from these pseudo-functions! Instead, they should be treated as returning `void`.

Differential Revision: D19158170

fbshipit-source-id: 68ed6b8a83baa0afb0cb6368ef494a12df202dbd