Delete dead comparator code.
Summary:
Fix the few places that were calling HPHP::same with a Variant and an int to pass an int64_t instead (they were all comparing the return value of HPHP::preg_match to either 0 or 1).
Fix the one C++ test that was calling HPHP::equal with ints.

"= delete" everything left in hphp/runtime/base/comparisons.h that takes an int (as opposed to an int64_t).

Delete the versions of tvEqual/Less/Greater that took a TV and an int and pased through to the versions that took a TV and an int64_t.

Delete the declared but not defined tvCompare that took a TV and an int.

Reviewed By: markw65

Differential Revision: D19152937

fbshipit-source-id: 3a8e3f3c59e5c59189446601d7b87e0bfbf6299d