Generalize `Memory_bounded_lru_cache` into memory-agnostic version called `Lru_cache`
Summary:
We'll switch from using a memory-bounded LRU cache to a size-bounded one.

There is a version of a `SharedMem` cache which is an LRU cache. However, it's always global; it offers a different API than what we need; it would require taking a dependency on the shared memory module; Zoncolan asked that we avoid mucking with those caches in case it causes a performance regression.

We've also instrumented our cache the way we like, so let's keep it.

Reviewed By: ljw1004

Differential Revision: D19149595

fbshipit-source-id: 800e330c7e78c157e0184d1d98bc1ab68f1173b8