Cap local-memory decl cache by number of decls, not by size
Summary:
`Obj.words_reachable` greatly overestimates the size of the entries in the cache now that we store the `Provider_backend` in the `Decl_env`. The `Decl_env` is closed over by the decl linearization process, and consequently it's reachable by the garbage collector, so every class decl contains a reference to the `Decl_env`. That means that every entry in our cache has a reference to the cache itself, which naturally blows out the cache for using "too much" memory (even though the memory is shared across all decls).

In practice, decls are so small that we can store many more than we should need and come out well under the 250 MB memory budget. This commit switches the decl cache to be a regular LRU cache, which mitigates most of the performance regression in the linked task.

Reviewed By: ljw1004

Differential Revision: D19149594

fbshipit-source-id: d486cdf3a8f886f77441487a5e6503eabfcc2bce