Add APIs to `Provider_utils` to compute TASTs without computing errors
Summary:
This happens often in IDE use-cases, where we only care about the TAST and not the errors. In particular, TAST checks can take a long time to run, and they don't contribute to the resulting TAST.

This commit doesn't actually change it so that TAST checks are skipped; that's in a future commit.

Reviewed By: ljw1004

Differential Revision: D19149593

fbshipit-source-id: 85788816bf654a6edf017344679ab96fc354bc6d