Fix order of translations
Summary:
sortTrans doesn't actually sort the translations; I assume thats
because we print them in the order they completed, which is going to
be roughly address order anyway. But it isn't /quite/ address order,
because we translate in parallel.

Lets actually sort them by address to get rid of the instability, so
that multiple jumpstart runs produce the same dump.

Reviewed By: paulbiss

Differential Revision: D19161568

fbshipit-source-id: bf139e5bcda2818918fa93f343d695401902e134