Eliminate (Dict)?AddElem(Int|Str)Key
Summary:
These ops do the same thing as ArraySet / DictSet, except that they also dec-ref the key. We can move the dec-ref of the key outside the op. In most cases, the string keys are static so doing so will actually save instructions.

By reusing ArraySet and DictSet, we also take advantage of the "move values" optimization. Overall, this change should be a small instructions win - small because these helpers are already quite cold (~0.1-0.2% of CPU time in total right now).

Reviewed By: mofarrell

Differential Revision: D19163753

fbshipit-source-id: de96e043d3a111d72fca27b464f6f1b14368ae8f