saved_state_loader support "env" in Utils.exec
Summary:
This diff adds support for environment to Utils.exec.

Why? Because a bunch of tools let their outputs be customized by environment variables. One such tool is `hg`, for which we want to use the environment HGPLAIN=1, to ensure we can scrape its output.

(I want telemetry to include the current commit hash, which will be obtained via Utils.exec "hg".)

I've added it as `Process_types.environment` rather than `string array`. That's because there's often confusion about whether a string array env is supposed to augment or replace the environment.

Reviewed By: 2BitSalute

Differential Revision: D19181235

fbshipit-source-id: d3d378e64067fd3d3b3fcf5cdbdfa92019c454e2