Restructure simplify_subtype
Summary:
This is a restructure of simplify_subtype. The goal is to make it clearer, since the current pattern matches we do make it hard to follow the code. The main structure goes as follows:

I've introduced a helper function in simplify_subtype_i, called `default_subtype` (open to suggestions for other names). This is meant to encode all the subtyping rules that are independent of the super type. For instance, the union as a subtype rule:
```
t1 <: t
t2 <: t
--------------
(t1 | t2) <: t
```

If there are no rules present then we return `invalid ()`.

There is then a separate match expression on `ety_super`. Each case on `ety_super` then matches against `ety_sub` on their own. Each super type can then encode the additional typing rules that are available knowing the super type. If there are no rules to apply, you call `default_subtype env` which will fallback to the default set of typing rules independent of the super type.

If ordering is important for the rules that apply, it can be done locally. For most types, it isn't important and this restructing optimizes for that case.

In this diff I convert some of the more difficult cases into this style to prove it works. Once it was in this style it was also possible to clean up some subsequent code as well.

Open to feedback, though I do believe this is a noticeable improvement to the current state and will make my implementation of like type coercion more clear.

Reviewed By: andrewjkennedy

Differential Revision: D19083834

fbshipit-source-id: 015819607bd74d38ba96b430eb32a33030d82d4e