Make Allocator API safe
Summary:
I chose to provide an unsafe API here in {D18008608} as a compromise. I wanted to borrow the allocator mutably to avoid paying for Cells in the allocator, but I didn't want to force consumers of Allocators to put all block fields on the stack before allocating the block (as implemented in {D17864962} and discussed in the comments). I was not able to come up with a design which accomplished all of 1) borrow the allocator mutably, 2) don't stack-allocate fields, and 3) safety. In D18008608, I sacrificed safety in exchange for the other two.

I probably overestimated the cost of Cells in the allocator (and the cost of bounds-checks for field writes). If so, we could give up item 1 instead of item 3. This diff does this by restoring the BlockBuilder type, representing an allocated, uninitialized block (i.e., a slice of word-length fields or binary data). It replaces the raw pointer previously returned by allocators. BlockBuilder cannot write to its own fields (since the ocamlpool allocator has its own logic it wants to run upon field writes), so we keep the Allocator::set_field API to allow allocators to define their own safe abstractions around field-writing behavior.

Reviewed By: 2BitSalute

Differential Revision: D19186747

fbshipit-source-id: 3871342c4e15f597c4c56438b09840b95d557eef