Mark Pool::new and to_ocaml unsafe
Summary: Any function which calls into the `ocamlpool` library may interact with the OCaml runtime, which is not thread-safe. I don't want to make `ocamlrep` APIs unsafe just because one `Allocator` (`Pool`) happens to perform unsafe operations, so instead I've marked `Pool::new` unsafe, with the requirement that the caller ensure that no other thread interacts with the OCaml runtime or ocamlpool while the Pool is alive. The `to_ocaml` function, which creates a `Pool`, has the same requirement.

Reviewed By: 2BitSalute

Differential Revision: D19199841

fbshipit-source-id: 209b9bae03b0ada5d6f21603174b92042026acae