Do not mark Value::to_bits as unsafe
Summary:
I marked this method unsafe because it is easy to cause a use-after-free using it and FFI with OCaml, and I wanted to call attention to the care that must be taken.

However, it's easy to cause problems like this in lots of other ways, and the language does not prevent safe code from producing raw pointers in other situations (for instance, it's safe to cast a reference to a raw pointer). This method does not violate memory safety on the Rust side--dereferencing the `usize` requires either unsafe code or FFI with another language which performs the dereference. Considering this, I'm not sure that there is a good reason to keep this method `unsafe`.

Reviewed By: dabek

Differential Revision: D19187735

fbshipit-source-id: 9e81e73d60cc24810d0e1fa761380f587eed94f1