Fix collection of implementation dependencies
Summary:
Consider the following example:
```
interface I {
  public function f(): void;
}

abstract class A {
  public function f(): void {}
}

final class B extends A implements I {}
```
Here `B` implements `I` by way of extending `A` whereas `A` itself is not declared as implementing `I`.  Suppose that `I::f` is a dependency of the target function.  (For a complete example, see a new test case below.)  We then check which implementations of `I::f` need to be pulled into the extracted code.  In the above example, `I` has only one descendant `B`, but `B::f` is *inherited* from `A`, i.e., its origin is `A` and is different from `B`, and currently we don't add implementations of such methods as dependencies, which may lead to type errors.  Instead, we should add `A::f` as a dependency, i.e., for each method implementation we need to add the pair of the method origin and the method name as a dependency, and ditto for static methods and type constants.

Reviewed By: andrewjkennedy

Differential Revision: D19194266

fbshipit-source-id: 3c790939fa4621c3dfe8b52c88c653018dab802e