Eliminate SetIntMoveRaw helper
Summary:
We can implement SetIntMove using MutableOpInt. We want to exit early when we hit the likely case where 0 <= k < size; we do by handling this case in the lambda and setting a flag that will force us to exit early.

An alternative is to put the extra handling (decref v; if escalated, dec-ref adIn) for both of the slow cases into each of the functions. I opted not to do that because it duplicates the source code and because it may duplicate the generated code.

Reviewed By: mofarrell

Differential Revision: D18694728

fbshipit-source-id: bcf124fa4d30d09f2e67a9a9b5378919bb74a4d0