Fix refcount-opts may-alias logic
Summary:
Our logic is too optimistic about annihilating IncRefs and DecRefs for asets that may-alias. For example, suppose asets 0 and 1 both have a lower bound of 2(1) (that is, a lower bound of 2, with 1 supported and 1 unsupported count). If these sets may-alias, the unsupported refcounts may be the other asets' supported refcounts, and we can convert two DecRefs incorrectly into DecRefNZs.

The fundamental problem here is that some unsupported references come from untracked memory locations, but others come from uses of values with a lower-bound of 0. We only clear unsupported references when we see memory operations on unsupported locations, so the second kind of unsupported references (as markw65 calls them, "assumed references") are too optimistic. This diff is a far-too-conservative fix for this problem: it essentially treats all unsupported references as assumed references, clearing them when any may-alias-set is dec-ref-ed. There are two orthogonal ways to make it less conservative:
1. As with the unbalanced dec-refs, we can keep a list of balanced dec-refs around and apply them only when observed.
2. We can split the unsupported_refs counter into untracked_refs and assumed_refs counters, and when applying a balanced dec-ref, only clear assumed_refs.

This change doesn't look like a regression, though, so I'll leave these two options as follow-ups.

Reviewed By: markw65

Differential Revision: D19271089

fbshipit-source-id: e5902906984213e82bee267ff31272bd6364c045