Use GADT for decl cache key-value pairs
Summary:
Use a GADT for a type-safe heterogeneous cache.

This migration exposes a real typing issue, where I was getting out an `'a`, but I thought it was an `'a option`. It happened to work by accident. GADTs are hard to understand but this makes it apparent to me that there is value in reducing the occurrences of `Obj` functions.

Reviewed By: ljw1004

Differential Revision: D19153590

fbshipit-source-id: 688b00bae6a051198632cfab681d126899be53dd